# 8272138: ZGC: Adopt relaxed ordering for self-healing