@weixlu 2021-08-09
@fisk Thanks for your comment. Yes, there is no need for marking to happen before self healing. But when an object needs to be copied to to-space in relocation phase, I guess self healing should not happen before copy. Otherwise the other thread may get a healed reference which hasn't finished copy yet. Since this pull request (https://github.com/openjdk/jdk/pull/4763) adopts release ordering for forwarding table access, maybe we need the release semantic here in self healing to prevent the reorder.