@tanghaoth90 2021-08-10
@fisk Hi, Eric. We are wondering if one thread loading a healed pointer can observe the corresponding copy has not finished yet. Assuming relaxed ordering for `cas_self_heal`, both Thread A and Thread B are loading the same reference.

**Thread A**:  `load obj.fld; // will relocate the object referenced by obj.fld`
thread A will do the following:
```
1    copy();
2    cas_forwarding_table(); // release
3    cas_self_heal(); // relaxed
```

**Thread B**:  `load obj.fld; // load the same reference`
thread B may obverses the following reordering of **thread A**:
```
3    cas_self_heal(); // relaxed
1    copy();
2    cas_forwarding_table(); // release
```
To our knowledge, release ordering in _line 2_ does not prevent _line 3_ to be reordering before _line 1_, which indicates the release in the forwarding table is not enough. Perhaps we need to add acquire ordering to _line 2_ or add release ordering to _line 3_.

In another way, as @weixlu said,
> Instead, it maybe serves as membar to block all the CASes afterwards.

relaxed ordering in _line 2_ along with release ordering in _line 3_ can indeed ensure thread B always observes the object copy.

Looking forward to your advice.